fix(functional): address code review bugs in perf optimizations
- Fix non-deterministic closest-source selection in generate_inverse_distortion_map:
  when multiple source pixels compete for the same destination cell, sort candidates
  descending by dist before writing so the minimum-dist source wins

- Fix float64 promotions leaking into float32 hot paths:
  - slic centers: .astype(np.float64) -> np.float32
  - slic inv_grid_step_sq: Python float literal -> np.float32(...)
  - create_piecewise_affine_maps weights: 1.0/(dist+1e-8) -> np.float32 constants

- Fix O(H*W*K) memory in create_piecewise_affine_maps: replace single
  (H,W,K) weight tensor with per-control-point accumulation loop,
  keeping memory O(H*W) — avoids >1 GB allocation for large grids

- Add test_closest_source_wins_on_collision to cover the collision fix

Made-with: Cursor

+++ albumentations/augmentations/pixel/functional.py
@@ -2509,13 +2509,13 @@ def slic(
     x_range = np.arange(grid_step // 2, width, grid_step)
     y_range = np.arange(grid_step // 2, height, grid_step)
     xx_grid, yy_grid = np.meshgrid(x_range, y_range)
-    centers = np.column_stack([xx_grid.ravel(), yy_grid.ravel()]).astype(np.float64)
+    centers = np.column_stack([xx_grid.ravel(), yy_grid.ravel()]).astype(np.float32)
 
     # Initialize labels and distances
     labels = np.full((height, width), -1, dtype=np.int32)
     distances = np.full((height, width), np.inf, dtype=np.float32)
 
-    inv_grid_step_sq = 1.0 / (grid_step * grid_step)
+    inv_grid_step_sq = np.float32(1.0 / (grid_step * grid_step))
 
     for _ in range(max_iterations):
         for i, center in enumerate(centers):
